refactor: entrypoint updated

+++ entrypoint.sh
@@ -1,8 +1,12 @@
 if [ ! -f .env ]; then
-  echo ".env not found, creating from .env.example"
-  cp .env.example .env
+  echo ".env not found in container, creating from .env.example"
+  if [ -f .env.example ]; then
+    cp .env.example .env
+  else
+    echo ".env.example not found either! The application might fail."
+  fi
 else
   echo ".env exists, skipping creation"
 fi
 
-exec "$@"
+exec uvicorn app.main:app --host 0.0.0.0 --port 8000
